+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -93,5 +93,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-05-23T11:32:36.222Z")
+    created: Optional[datetime] = parse_datetime("2025-05-24T19:14:42.496Z")
     r"""The date and time the resource was created"""