+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -78,5 +78,5 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-05-23T11:32:33.441Z")
+    updated: Optional[datetime] = parse_datetime("2025-05-24T19:14:39.719Z")
     r"""The date and time the resource was last updated"""