+++ packages/orq-rc/RELEASES.md
@@ -1479,3 +1479,13 @@ Based on:
 - [python v3.6.0-rc.33] packages/orq-rc
 ### Releases
 - [PyPI v3.6.0-rc.33] https://pypi.org/project/orq-ai-sdk/3.6.0-rc.33 - packages/orq-rc
+
+## 2025-05-24 19:16:07
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.551.0 (2.610.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.6.0-rc.34] packages/orq-rc
+### Releases
+- [PyPI v3.6.0-rc.34] https://pypi.org/project/orq-ai-sdk/3.6.0-rc.34 - packages/orq-rc